Remove derivation of QI and QL from other fields; read from asm collection
This update differs from GEOS-FP. While QI and QL are in the asm
collection of GEOS-FP raw data they are not read or used. Instead QI and
QL are derived from other met-fields. For GEOS-IT we will read the
values stored in the raw data.

This update requires updating the data structure storing asm collection
data two be two different arrays, one for fields going into the
A3Cld processed file collection and the other for fields going into the
A3Dyn processed file collection.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>